Break out impls, remove Varint
