+++ source/postcard-schema/src/impls/builtins_alloc.rs
@@ -1,0 +1,38 @@
+use crate::{
+    Schema,
+    schema::{NamedType, SdmTy},
+};
+
+
+extern crate alloc;
+
+impl<T: Schema> Schema for alloc::vec::Vec<T> {
+    const SCHEMA: &'static NamedType = &NamedType {
+        name: "Vec<T>",
+        ty: &SdmTy::Seq(T::SCHEMA),
+    };
+}
+
+impl Schema for alloc::string::String {
+    const SCHEMA: &'static NamedType = &NamedType {
+        name: "String",
+        ty: &SdmTy::String,
+    };
+}
+
+impl<K: Schema, V: Schema> Schema for alloc::collections::BTreeMap<K, V> {
+    const SCHEMA: &'static NamedType = &NamedType {
+        name: "BTreeMap<K, V>",
+        ty: &SdmTy::Map {
+            key: K::SCHEMA,
+            val: V::SCHEMA,
+        },
+    };
+}
+
+impl<K: Schema> Schema for alloc::collections::BTreeSet<K> {
+    const SCHEMA: &'static NamedType = &NamedType {
+        name: "BTreeSet<K>",
+        ty: &SdmTy::Seq(K::SCHEMA),
+    };
+}
