+++ source/postcard-schema/src/impls/builtins_nostd.rs
@@ -1,0 +1,98 @@
+use crate::{
+    Schema,
+    schema::{NamedType, NamedVariant, SdmTy},
+};
+use core::num::{
+    NonZeroI128, NonZeroI16, NonZeroI32, NonZeroI64, NonZeroI8, NonZeroU128, NonZeroU16,
+    NonZeroU32, NonZeroU64, NonZeroU8,
+};
+
+macro_rules! impl_schema {
+    ($($t:ty: $sdm:expr),*) => {
+        $(
+            impl Schema for $t {
+                const SCHEMA: &'static NamedType = &NamedType {
+                    name: stringify!($t),
+                    ty: &$sdm,
+                };
+            }
+        )*
+    };
+    (tuple => [$(($($generic:ident),*)),*]) => {
+        $(
+            impl<$($generic: Schema),*> Schema for ($($generic,)*) {
+                const SCHEMA: &'static NamedType = &NamedType {
+                    name: stringify!(($($generic,)*)),
+                    ty: &SdmTy::Tuple(&[$($generic::SCHEMA),*]),
+                };
+            }
+        )*
+    };
+}
+
+impl_schema![
+    u8: SdmTy::U8,
+    NonZeroU8: SdmTy::U8,
+    i8: SdmTy::I8,
+    NonZeroI8: SdmTy::I8,
+    bool: SdmTy::Bool,
+    f32: SdmTy::F32,
+    f64: SdmTy::F64,
+    char: SdmTy::Char,
+    str: SdmTy::String,
+    (): SdmTy::Unit,
+    i16: SdmTy::I16, i32: SdmTy::I32, i64: SdmTy::I64, i128: SdmTy::I128,
+    u16: SdmTy::U16, u32: SdmTy::U32, u64: SdmTy::U64, u128: SdmTy::U128,
+    NonZeroI16: SdmTy::I16, NonZeroI32: SdmTy::I32,
+    NonZeroI64: SdmTy::I64, NonZeroI128: SdmTy::I128,
+    NonZeroU16: SdmTy::U16, NonZeroU32: SdmTy::U32,
+    NonZeroU64: SdmTy::U64, NonZeroU128: SdmTy::U128
+];
+
+impl_schema!(tuple => [
+    (A),
+    (A, B),
+    (A, B, C),
+    (A, B, C, D),
+    (A, B, C, D, E),
+    (A, B, C, D, E, F)
+]);
+
+impl<T: Schema> Schema for Option<T> {
+    const SCHEMA: &'static NamedType = &NamedType {
+        name: "Option<T>",
+        ty: &SdmTy::Option(T::SCHEMA),
+    };
+}
+impl<T: Schema, E: Schema> Schema for Result<T, E> {
+    const SCHEMA: &'static NamedType = &NamedType {
+        name: "Result<T, E>",
+        ty: &SdmTy::Enum(&[
+            &NamedVariant {
+                name: "Ok",
+                ty: &SdmTy::TupleVariant(&[T::SCHEMA]),
+            },
+            &NamedVariant {
+                name: "Err",
+                ty: &SdmTy::TupleVariant(&[E::SCHEMA]),
+            },
+        ]),
+    };
+}
+
+impl<T: Schema + ?Sized> Schema for &'_ T {
+    const SCHEMA: &'static NamedType = T::SCHEMA;
+}
+
+impl<T: Schema> Schema for [T] {
+    const SCHEMA: &'static NamedType = &NamedType {
+        name: "&[T]",
+        ty: &SdmTy::Seq(T::SCHEMA),
+    };
+}
+impl<T: Schema, const N: usize> Schema for [T; N] {
+    const SCHEMA: &'static NamedType = &NamedType {
+        name: "[T; N]",
+        ty: &SdmTy::Tuple(&[T::SCHEMA; N]),
+    };
+}
