+++ source/postcard-schema/src/impls/chrono_v0_4.rs
@@ -1,0 +1,11 @@
+use crate::{
+    Schema,
+    schema::NamedType,
+};
+
+impl<Tz: chrono::TimeZone> Schema for chrono::DateTime<Tz> {
+    const SCHEMA: &'static NamedType = &NamedType {
+        name: "DateTime",
+        ty: <&str>::SCHEMA.ty,
+    };
+}
