+++ source/postcard-schema/src/impls/uuid_v1_0.rs
@@ -1,0 +1,11 @@
+use crate::{
+    Schema,
+    schema::{NamedType, SdmTy},
+};
+
+impl Schema for uuid::Uuid {
+    const SCHEMA: &'static NamedType = &NamedType {
+        name: "Uuid",
+        ty: &SdmTy::ByteArray,
+    };
+}
